pyproject.toml
[project]
name = "prompt-optimizaton"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langchain (>=0.3.27,<0.4.0)",
    "langgraph (>=0.6.7,<0.7.0)",
    "langchain-openai (>=0.3.33,<0.4.0)",
    "numpy (>=2.3.3,<3.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
